+++ Civi/ActionProvider/Action/Contribution/LinkContributionToMembership.php
@@ -1,0 +1,57 @@
+<?php
+/**
+ * @author Jaap Jansma <jaap.jansma@civicoop.org>
+ * @license AGPL-3.0
+ */
+
+namespace Civi\SepaMembership\ActionProvider\Action;
+
+use Civi\ActionProvider\Action\AbstractAction;
+use Civi\ActionProvider\Parameter\ParameterBagInterface;
+use Civi\ActionProvider\Parameter\Specification;
+use Civi\ActionProvider\Parameter\SpecificationBag;
+
+use CRM_Sepamembership_ExtensionUtil as E;
+
+class LinkContributionToMembership extends AbstractAction {
+
+  /**
+   * Run the action
+   *
+   * @param ParameterBagInterface $parameters
+   *   The parameters to this action.
+   * @param ParameterBagInterface $output
+   *   The parameters this action can send back
+   *
+   * @return void
+   */
+  protected function doAction(ParameterBagInterface $parameters, ParameterBagInterface $output) {
+    $apiParams['contribution_id'] = $parameters->getParameter('contribution_id');
+    $apiParams['membership_id'] = $parameters->getParameter('membership_id');
+    civicrm_api3('MembershipPayment', 'create', $apiParams);
+  }
+
+  /**
+   * Returns the specification of the configuration options for the actual
+   * action.
+   *
+   * @return SpecificationBag
+   */
+  public function getConfigurationSpecification() {
+    return new SpecificationBag(array());
+  }
+
+  /**
+   * Returns the specification of the parameters of the actual action.
+   *
+   * @return SpecificationBag
+   */
+  public function getParameterSpecification() {
+    return new SpecificationBag(array(
+      new Specification('membership_id', 'Integer', E::ts('Membership ID'), true),
+      new Specification('contribution_id', 'Integer', E::ts('Contribution ID'), true),
+    ));
+  }
+
+
+}